% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_functions.R
\name{ADMM}
\alias{ADMM}
\title{Proximal ADMM algorithm}
\usage{
ADMM(x0, y0, prox_f, prox_g, step_size, max_iter, eps_abs, eps_rel, ...)
}
\arguments{
\item{x0}{initial value of primal variable (c(x0, z0))}

\item{y0}{initial value of dual variable}

\item{prox_f}{function for computing the proximal mapping of the objective}

\item{prox_g}{function for computing the proximal mapping of the constraint}

\item{step_size}{step parameter for prox_f}

\item{max_iter}{maximum number of iterations}

\item{eps_abs}{absolute epsilon for stopping criteria}

\item{eps_rel}{relative epsilon for stopping criteria}
}
\description{
\code{ADMM} Runs the proximal admm algorithm for minimizing
f1(x) + f2(z) subject to z = Dx.
}
