% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_tau.R
\name{best_tau}
\alias{best_tau}
\title{Find the best tau parameter for exponential smoothing detrending.}
\usage{
best_tau(img, cutoff = 0.05, seed = NULL, parallel = FALSE)
}
\arguments{
\item{img}{A 3-dimensional array of integers representing the image series.
\code{img[i, j, k]} is the pixel at \code{x = i}, \code{y = j} in frame \code{k} of the image
series.}

\item{cutoff}{When exponential smoothing, neglect weights which are less than
\code{cutoff * central weight}. It is fine to leave this parameter at the
default level.}

\item{seed}{A seed for the random number generation needed in this function.
If you do not set a seed, one will be chosen randomly.}

\item{parallel}{Do you want this function to run on multiple cores? If not,
use \code{parallel = FALSE} (the default). To use \code{n} cores, use \code{parallel = n}
or to use all available cores use \code{parallel = TRUE}.}
}
\value{
If no detrend is necessary, this function returns \code{NA}. If a detrend
is required, this function returns a natural number which is the ideal
\code{tau} parameter for exponential smoothing detrending.
}
\description{
Use Nolan's algorithm to find the ideal tau parameter for exponential
smoothing detrending.
}
\examples{
\dontrun{
## These examples are not run on CRAN because they take too long.
## You should still try them for yourself.
img <- read_tif(system.file('extdata', 'bleached.tif', package = 'detrendr'),
                n_ch = 1)
best_tau(img, seed = 0, parallel = 2)
}
}
\references{
Rory Nolan, Luis A. J. Alvarez, Jonathan Elegheert, Maro
Iliopoulou, G. Maria Jakobsdottir, Marina Rodriguez-Muñoz, A. Radu
Aricescu, Sergi Padilla-Parra; nandb—number and brightness in R with a
novel automatic detrending algorithm, Bioinformatics,
https://doi.org/10.1093/bioinformatics/btx434.
}
