% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_loss}
\alias{check_loss}
\title{Check loss function
\code{check_loss} Evaluates check loss function}
\usage{
check_loss(r, tau)
}
\arguments{
\item{r}{vector of residuals}

\item{tau}{quantile level must be in [0,1]}
}
\description{
Check loss function
\code{check_loss} Evaluates check loss function
}
