% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chol_solve}
\alias{chol_solve}
\title{Banded Cholesky Solve}
\usage{
chol_solve(cholM, b, k, upper = TRUE)
}
\arguments{
\item{cholM}{sparse banded cholseky decomposition of discrete derivative
matrix of order k}

\item{b}{dense solution vector}

\item{k}{order of discrete derivative matrix}

\item{upper}{boolean indicator of whether cholM is upper or lower triangular}
}
\description{
\code{chol_solve} Solves a linear system cholM%*%x=b
when cholM is a sparse banded cholesky.
}
