% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_ADMM.R
\name{consensus_ADMM}
\alias{consensus_ADMM}
\title{Consensus ADMM for overlapping windows}
\usage{
consensus_ADMM(y, tau, lambda, k, rho, window_size, overlap, max_iter,
  eps = 0.01, update = 10)
}
\arguments{
\item{y}{observed data}

\item{tau}{quantile levels at which to evaluate trend}

\item{lambda}{smoothing penalty parameter}

\item{k}{order of differencing}

\item{rho}{step size for ADMM}

\item{window_size}{size of windows to use}

\item{overlap}{integer length of overlap between windows}

\item{max_iter}{Maximum number of iterations}
}
\description{
\code{consensus ADMM} Uses ADMM to smooth overlapping windows
}
\examples{
n <- 100
x <- seq(1, n, 1)
y <- sin(x*2*pi/n) + rnorm(n, 0, .4)
lambda <- 10
k <- 3
tau <- c(0.05)
overlap <- 20
rho <- 1
result <- consensus_ADMM(y, tau, lambda, k, rho, overlap, 300)
y_n <- length(y)
window_size <- floor((y_n+overlap)/2)
plot(result$phiBar~x, type="l")
points(result$phi1~x[1:window_size])
points(result$phi2~x[(y_n-window_size+1):y_n], col="blue")
plot(y~x)
lines(result$theta)
plot(result$primal_norm)
plot(result$dual_norm)
}
