% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{display}
\alias{display}
\title{Basic image display.}
\usage{
display(img, ...)
}
\arguments{
\item{img}{A matrix.}

\item{...}{Arguments passed to \code{\link[fields:image.plot]{fields::image.plot()}}. These arguments should
be named.}
}
\description{
Display an image that has been read in by \code{\link[=read_tif]{read_tif()}} as it would look in
ImageJ. This function wraps \code{\link[fields:image.plot]{fields::image.plot()}}.
}
\examples{
img <- read_tif(system.file('extdata', 'bleached.tif', package = 'detrendr'),
                n_ch = 1)
display(img[, , 1])

}
