% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simluate_peaks.R
\name{generate_peaks}
\alias{generate_peaks}
\title{Simulate data with peaks and baseline}
\usage{
generate_peaks(n)
}
\arguments{
\item{n}{length of series to generate}
}
\description{
Simulate data with peaks and baseline
}
