% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_constraint_mat}
\alias{get_constraint_mat}
\title{Get constraint matrix for linear program}
\usage{
get_constraint_mat(D, nT)
}
\arguments{
\item{D}{discrete difference matrix}

\item{nT}{number of quantiles being estimated}
}
\description{
\code{get_constraint_mat} Returns the constraint matrix
}
