% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_criteria.R
\name{get_criteria}
\alias{get_criteria}
\title{\code{get_criteria}}
\usage{
get_criteria(criteria, f_trend, y, tau, D = NULL, df_tol = 1e-09,
  gamma = 1, validID = NULL, yValid = NULL)
}
\arguments{
\item{criteria}{label of criteria to be used, must be one of "eBIC", "SIC", "valid"}

\item{f_trend}{fitted quantile trend(s)}

\item{tau}{quantile vector}

\item{D}{discrete differencing matrix (for SIC and eBIC)}

\item{df_tol}{tolerance for determining degrees of freedom (D%*%theta > df_tol)}

\item{gamma}{parameter for eBIC}

\item{validID}{index of data to be used for validation (for valid method)}

\item{yValid}{validation data (for valid method)}

\item{observed}{data vector}
}
\description{
\code{get_criteria}
}
