% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{get_D1}
\alias{get_D1}
\title{First order difference matrix}
\usage{
get_D1(n)
}
\arguments{
\item{n}{length of input}
}
\description{
\code{get_D1} computes the discrete derivative matrix.
}
\examples{
n <- 5
D1 <- get_D1(n)
}
