% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_Dk}
\alias{get_Dk}
\title{kth order sparse difference matrix}
\usage{
get_Dk(n, k)
}
\arguments{
\item{n}{length of input}

\item{k}{order of the derivative}
}
\description{
\code{get_Dkn} computes the sparse discrete kth derivative matrix
}
