% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_windows.R
\name{get_eta}
\alias{get_eta}
\title{Update dual variable (overlap)}
\usage{
get_eta(phi, y, k)
}
\arguments{
\item{phi}{matrix of y-theta}

\item{y}{vector of observations}

\item{k}{order of differencing matrix}
}
\description{
\code{get_eta}
}
