% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metric.R
\name{get_metric}
\alias{get_metric}
\title{Get evaluation metrics for classification}
\usage{
get_metric(trend, y, signal, threshold, metric)
}
\arguments{
\item{trend}{estimated trend vector}

\item{y}{vector of observations}

\item{signal}{vector of true signal classification}

\item{threshold}{value for threshold defining signal}

\item{metric}{choice of "missclass", "recall", "CAA", "precision" or "F1"}
}
\description{
\code{get_metric} Returns classification metric requested
}
