% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Get linear program arguments for gurobi or glpk solver}
\usage{
get_model(y, tau, lambda, k)
}
\arguments{
\item{y}{observed data, should be equally spaced, may contain NA}

\item{tau}{vector of quantile levels at which to evaluate trend}

\item{lambda}{vector of penalty parameters controlling smoothness}

\item{k}{order of differencing}
}
\description{
\code{get_model} Returns the linear program solver arguments
}
