% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_updates.R
\name{get_model_abs}
\alias{get_model_abs}
\title{Get model with ADMM step parameter and dual and consensus variables}
\usage{
get_model_abs(model, z, rho, nT)
}
\arguments{
\item{model}{Model object}

\item{z}{consensus variable}

\item{rho}{ADMM step parameter}

\item{nT}{number of quantile levels}
}
\description{
\code{get_model_abs}
}
