% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_objective}
\alias{get_objective}
\title{Get objective linear program}
\usage{
get_objective(tau, lambda, n, m, missInd)
}
\arguments{
\item{tau}{quantiles being estimated}

\item{lambda}{smoothing parameters, should be same length as tau}

\item{n}{length of data}

\item{m}{n-k where k is order of differencing matrix}

\item{missInd}{output of which(is.na(y))}
}
\description{
\code{get_objective} Returns objective coefficent vector
}
