% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_windows.R
\name{get_phiBar}
\alias{get_phiBar}
\title{Update consensus variable (overlap)}
\usage{
get_phiBar(phiBar_list, windows)
}
\arguments{
\item{phiBar_list}{List of consensus variable}

\item{windows}{Matrix indicating window group}
}
\description{
\code{get_phiBar}
}
