% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_trend}
\alias{get_trend}
\title{Get quantile trends}
\usage{
get_trend(y, tau, lambda, k)
}
\arguments{
\item{y}{observed data, should be equally spaced, may contain NA}

\item{tau}{quantile levels at which to evaluate trend}

\item{lambda}{penalty paramter controlling smoothness}

\item{k}{order of differencing}
}
\description{
\code{get_trend} Returns the quantile trend matrix
}
\examples{
require(Matrix)
n <- 100
x <- seq(1, n, 1)
y <- sin(x*2*pi/n) + rnorm(n, 0, .4)
lambda <- 10
k <- 3
tau <- c(0.05, .2)
trend <- get_trend(y, tau, lambda, k)
plot(trend[,1]~x, type="l")
}
