% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend_spingarn.R
\name{get_trend_spingarn}
\alias{get_trend_spingarn}
\title{Get quantile trend using Spingarn's Method}
\usage{
get_trend_spingarn(y, tau, lambda, k, numIter = 10000)
}
\arguments{
\item{y}{observed data, should be equally spaced, may contain NA}

\item{tau}{quantile level at which to evaluate trend}

\item{lambda}{penalty paramter controlling smoothness}

\item{k}{order of differencing}

\item{numIter}{number of iterations to run algorithm}
}
\description{
\code{get_trend_spingarn} Returns the quantile trend vector
}
