% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend_windows.R
\name{get_trend_windows}
\alias{get_trend_windows}
\title{Consensus ADMM for overlapping windows}
\usage{
get_trend_windows(y, tau, lambda, k, window_size, overlap, max_iter,
  rho = 5, update = 10, use_gurobi = TRUE, eps_abs = 0.05,
  eps_rel = 0.001)
}
\arguments{
\item{y}{observed data}

\item{tau}{quantile levels at which to evaluate trend}

\item{lambda}{smoothing penalty parameter}

\item{k}{order of differencing}

\item{window_size}{size of windows to use}

\item{overlap}{integer length of overlap between windows}

\item{max_iter}{Maximum number of iterations}

\item{rho}{parameter for ADMM}

\item{update}{number of iterations at which to print residuals}

\item{use_gurobi}{TRUE if gurobi solver is installed and should be used}

\item{eps_abs}{absolute threshold for stopping criteria}

\item{eps_rel}{relative threshold for stopping criteria}
}
\description{
\code{consensus ADMM} Uses ADMM to smooth overlapping windows
}
\examples{
require(Matrix)
n <- 100
x <- seq(1, n, 1)
y <- sin(x*2*pi/n) + rnorm(n, 0, .4)
lambda <- 10
k <- 3
y_n <- length(y)
overlap <- 20
window_size <- round((y_n+overlap)/2)
tau <- c(0.05, .2)
max_iter <- 20
trend <- get_trend_windows(y, tau, lambda, k, window_size, overlap, max_iter)
plot(trend[,1]~x, type="l")
}
