% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_windows.R
\name{get_windows}
\alias{get_windows}
\title{Estimate trend using rolling windows}
\usage{
get_windows(y, time, k, tau, lambda, window_size, overlap)
}
\arguments{
\item{y}{observed data}

\item{time}{timestamp for observations}

\item{k}{order of differencing}

\item{tau}{quantile levels at which to evaluate trend}

\item{lambda}{smoothing penalty parameter}

\item{window_size}{integer length of window}

\item{overlap}{integer length of overlap between windows}
}
\description{
\code{get_windows} Estimates trend in rolling windows
}
