% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi_trend.R
\name{gurobi_trend}
\alias{gurobi_trend}
\title{Quantile Regression with Trend Filtering (Gurobi Solver)}
\usage{
gurobi_trend(y, tau, lambda, k)
}
\arguments{
\item{y}{observed data}

\item{tau}{quantile levels at which to evaluate trend}

\item{lambda}{penalty paramter controlling smoothness}

\item{k}{order of differencing}
}
\description{
\code{gurobi_trend} Determines the quantile regression trend filtering solution
}
