% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_criteria.R
\name{H}
\alias{H}
\title{# Functions for choosing smoothing parameter}
\usage{
H(x)
}
\description{
#' Evaluate checkloss function
#'
#' \code{checkloss}
#'
#' @param e argument of checkloss function 
#' @param tau quantile to be used
#' @export
checkloss <- function(e, tau){
  if (ncol(e) != length(tau)){
    stop("Number of columns in y must be same as length of tau")
  }
  obj <- e
  for (i in 1:length(tau)){
    obj[,i] <- obj[,i]*tau[i]
    obj[e[,i] < 0,i] <- e[e[,i] < 0,i]*(tau[i]-1)
  }
  return(obj)
  
}
}
