% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_functions.R
\name{LADMM}
\alias{LADMM}
\title{Linearized ADMM algorithm}
\usage{
LADMM(x0, y0, A, prox_f, prox_g, step_f, step_g, max_iter, ...)
}
\arguments{
\item{x0}{initial value of primal variable}

\item{y0}{initial value of dual variable}

\item{prox_f}{function for computing the proximal mapping of f}

\item{prox_g}{function for computing the proximal mapping of g}

\item{step_f}{step parameter for prox_f with 0 < step_f < step_g/||A||^2}

\item{step_g}{step parameter for prox_g}

\item{max_iter}{maximum number of iterations}
}
\description{
\code{ADMM} Runs the Linearized ADMM algorithm for minimizing
f(x) + g(Ax)
}
