% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_cv.R
\name{lambda_cv}
\alias{lambda_cv}
\title{Choose smoothing parameter using crossvalidation}
\usage{
lambda_cv(y, tau, k, lambdaSeq = c(length(y)/10, length(y), length(y) *
  10), numFolds = 5, parallel = TRUE)
}
\arguments{
\item{y}{observed data}

\item{tau}{quantile levels at which to evaluate trend}

\item{k}{order of differencing}

\item{lambdaSeq}{smoothing penalty parameter options to compare}

\item{numFolds}{number of cross-validation folds to use}

\item{parallel}{TRUE/FALSE run in lambdas in parallel}
}
\description{
\code{lambda_cv} Selects smoothing parameter using cross validation
}
