% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_SIC.R
\name{lambda_SIC}
\alias{lambda_SIC}
\title{Choose smoothing parameter using SIC}
\usage{
lambda_SIC(y, tau, k, lambdaSeq = c(length(y)/10, length(y), length(y) *
  10), df_tol = 1e-09)
}
\arguments{
\item{y}{observed data}

\item{tau}{quantile levels at which to evaluate trend}

\item{k}{order of differencing}

\item{lambdaSeq}{smoothing penalty parameter options to compare}
}
\description{
\code{lambda_SIC} Selects smoothing parameter using Schwarz Information
}
