% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project_V}
\alias{project_V}
\title{Project onto subspace (updates values of theta and eta in place)

\code{project_V} projects (theta, eta) onto the subspace eta = D%*%theta}
\usage{
project_V(theta, eta, D, cholM)
}
\arguments{
\item{theta}{first input}

\item{eta}{second input}

\item{D}{differencing matrix}

\item{cholM}{upper triangular cholesky decomposition of  I + DtD}
}
