% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{project_V_R}
\alias{project_V_R}
\title{Project onto subspace}
\usage{
project_V_R(theta, eta, D, M)
}
\arguments{
\item{theta}{first input}

\item{eta}{second input}

\item{D}{differencing matrix}
}
\description{
\code{project_V} projects (theta, eta) onto the subspace eta = D%*%theta
}
\examples{
set.seed(12345)
k <- 3
n <- 1e2
D <- get_Dk(n, k)
theta <- rnorm(n)
eta <- D \%*\% theta
proj <- project_V(theta, eta, D)

m <- 1e3
angles <- double(m)
theta2 <- matrix(rnorm(n*m), n, m)
eta2 <- D\%*\%theta2
angles <- matrix(t(theta - proj$theta)\%*\%(theta2 - matrix(proj$theta)\%*\%matrix(1,1,m)))
angles <- angles + matrix(t(eta - proj$eta)\%*\%(eta2 - matrix(proj$eta)\%*\%matrix(1,1,m)))
summary(angles)
}
