% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox}
\alias{prox}
\title{Proximal mapping}
\usage{
prox(theta, eta, y, lambda, tau = 0.05, step = 1)
}
\arguments{
\item{theta}{input}

\item{eta}{input}

\item{y}{response}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{step}{step-size}
}
\description{
\code{prox} computes the block separable proximal mapping, changes theta
and eta in place
}
