% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_functions.R
\name{prox_f}
\alias{prox_f}
\title{Proximal mapping of quantile trend filtering objective}
\usage{
prox_f(x0, step_size, lambda, y, D, M, tau)
}
\arguments{
\item{x0}{input initial value of c(theta, eta)}

\item{step_size}{step parameter of prox}

\item{lambda}{penalty parameter for differences}

\item{y}{observed data}

\item{tau}{quantile level}
}
\description{
\code{prox_f} computes the proximal mapping of the quantile trend filtering objective
}
