% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{prox_f1}
\alias{prox_f1}
\title{Proximal mapping of f_1}
\usage{
prox_f1(theta, y, tau = 0.05, t = 1)
}
\arguments{
\item{theta}{input}

\item{y}{response}

\item{tau}{quantile parameter}

\item{t}{step-size}
}
\description{
\code{prox_f1} computes the proximal mapping of the average quantile loss
}
\examples{
set.seed(12345)
}
