% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{prox_f1_R}
\alias{prox_f1_R}
\title{Proximal mapping of f_1}
\usage{
prox_f1_R(theta, y, tau, step)
}
\arguments{
\item{theta}{input}

\item{y}{response}

\item{tau}{quantile parameter}

\item{t}{step-size
set.seed(12345)}
}
\description{
\code{prox_f1} computes the proximal mapping of the average quantile loss
}
