% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_f2}
\alias{prox_f2}
\title{Proximal mapping of f_2}
\usage{
prox_f2(eta, lambda, step = 1)
}
\arguments{
\item{eta}{input}

\item{lambda}{regularization parameter}

\item{step}{step-size}
}
\description{
\code{prox_f2} computes the proximal mapping of the L1 penalty
}
\examples{
set.seed(12345)
n <- 1e3
eta <- seq(-3, 3, length.out=n)
lambda <- 1
prox_out <- prox_f2(eta, lambda)
plot(eta, prox_out, type = 'l')
abline(0,1)
}
