% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_functions.R
\name{prox_g}
\alias{prox_g}
\title{Proximal mapping of indicator eta = D*theta}
\usage{
prox_g(x0, lambda, y, D, M, tau)
}
\arguments{
\item{x0}{input initial value of c(theta, eta)}

\item{D}{differencing matrix}

\item{M}{cholesky factorization of (crossprod(D) + I)}
}
\description{
\code{prox_g} Computes the proximal mapping of the constraint (projection)
}
