% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_quantile}
\alias{prox_quantile}
\title{Proximal Mapping}
\usage{
prox_quantile(w, tau, alpha)
}
\arguments{
\item{w}{input}

\item{tau}{quantile parameter}

\item{alpha}{scale parameter}
}
\description{
\code{prox_quantile} computes the proximal mapping of the check function.
}
\examples{
set.seed(12345)
n <- 1e3
w <- seq(-3, 3, length.out=n)
tau <- 0.5
alpha <- 2
prox_out <- prox_quantile(w, tau, alpha)
plot(w, prox_out, type='l', main=expression(paste(tau," = ")))

tau <- 0.05
alpha <- 2
prox_out <- prox_quantile(w, tau, alpha)
plot(w, prox_out, type='l', main=expression(paste(tau," = ")))
}
