% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_test}
\alias{prox_test}
\title{Proximal Mapping Test

\code{prox_test} computes the block separable proximal mapping.
Returns values of theta and eta.}
\usage{
prox_test(theta, eta, y, lambda, tau = 0.05, step = 1)
}
\arguments{
\item{theta}{input}

\item{eta}{input}

\item{y}{response}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{step}{step-size}
}
