% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_updates.R
\name{quad_update}
\alias{quad_update}
\title{Solve quadratic program step in consensus ADMM}
\usage{
quad_update(w, z, model, rho, nT)
}
\arguments{
\item{w}{value of dual variable}

\item{z}{consensus variable}

\item{rho}{step size for ADMM}

\item{nT}{number of quantiles}
}
\description{
\code{quad_update} Uses gurobi to solve quadratic program step
}
