% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_functions.R
\name{solve_model}
\alias{solve_model}
\title{Solves the model using chosen solver}
\usage{
solve_model(model, solver, y = NULL, trend = TRUE)
}
\arguments{
\item{model}{model object containing coefficients and constraints}

\item{solver}{solver to be used, options are "gurobi", "Rglpk", and "lpSolve"}

\item{y}{vector of observations, only needed if trend = TRUE}

\item{trend}{if TRUE returns trend, if FALSE returns residuals}
}
\description{
\code{solve_model} Returns the trend
}
