% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spingarn_multi_step}
\alias{spingarn_multi_step}
\title{Multiple steps of Spingarn's algorithm}
\usage{
spingarn_multi_step(theta, eta, y, D, cholM, lambda, tau = 0.05,
  step = 1, numberIter = 1, k = 3L)
}
\arguments{
\item{theta}{input 1}

\item{eta}{input 2}

\item{y}{response}

\item{D}{discrete differencing matrix}

\item{cholM}{cholesky of I + DtD}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{step}{step-size}

\item{numberIter}{number of iterations}

\item{k}{order of differencing}
}
\description{
\code{spingarn_multi_step}
}
