% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{spingarn_multi_step_R}
\alias{spingarn_multi_step_R}
\title{Multiple step of Spingarn's algorithm}
\usage{
spingarn_multi_step_R(theta, eta, y, D, M, lambda, tau = 0.05, t = 1,
  numberIter = 1)
}
\arguments{
\item{theta}{input 1}

\item{eta}{input 2}

\item{y}{response}

\item{D}{differencing matrix}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{t}{step-size}
}
\description{
\code{spingarn_one_step} computes a single round of Spingarn updates.
}
