% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spingarn_multistart}
\alias{spingarn_multistart}
\title{Spingarn's algorithm multiple initial values}
\usage{
spingarn_multistart(theta1, eta1, theta2, eta2, y, D, cholM, lambda,
  tau = 0.05, step = 1, numberIter = 1, k = 3L)
}
\arguments{
\item{y}{response}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{step}{step-size}

\item{numberIter}{number of iterations}

\item{k}{order of derivative}
}
\description{
\code{spingarn_multistart}
}
