% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{spingarn_one_step_R}
\alias{spingarn_one_step_R}
\title{One step of Spingarn's algorithm}
\usage{
spingarn_one_step_R(theta, eta, y, D, lambda, tau = 0.05, t = 1, M)
}
\arguments{
\item{theta}{input 1}

\item{eta}{input 2}

\item{y}{response}

\item{D}{differencing matrix}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{t}{step-size}
}
\description{
\code{spingarn_one_step} computes a single round of Spingarn updates.
}
\examples{
set.seed(12345)
n <- 1e2
x <- seq(1/n, 1, length.out=n)
f <- 2*(x + 2)^2 + 3*cos(3*pi*x)
tau <- 1e4
g <-100*exp(-tau*(x-0.5)^2)
y <- f + g + rnorm(n)
k <- 3
D <- get_Dk(n, k)
lambda <- 1
tau <- 0.01
t <- 1
theta <- y
eta <- matrix(D \%*\% theta)

max_iter <- 1e2
THx <- matrix(NA, n, max_iter)
for (iter in 1:max_iter) {
   one_step <- spingarn_one_step(theta, eta, y, D, lambda, tau=tau, t=t)
   theta <- one_step$theta
   eta <- one_step$eta
   THx[,iter] <- prox_f1(theta, y, tau)
}
theta_last <- prox_f1(theta, y, tau, t=t)
rerr <- double(max_iter-1)
for (i in 1:(max_iter-1)) {
  rerr[i] <- norm(as.matrix(THx[,i+1] - THx[,i]),'f')/(1 + norm(THx[,i,drop=FALSE], 'f'))
}
plot(x,f,type='l',col='blue', ylim=c(min(y),max(y)), lwd=3)
points(x,y,pch=16)
lines(x,theta_last,col='red', lwd=3)
}
