% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_Dk}
\alias{test_Dk}
\title{Function to test differencing matrix, returns single element of Dk(n)}
\usage{
test_Dk(n, k, row, col)
}
\arguments{
\item{n}{length of input}

\item{k}{order of the derivative}

\item{row}{row of element to return}

\item{col}{column of element to return}
}
