% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_updates.R
\name{update_consensus}
\alias{update_consensus}
\title{Update consensus variable (overlap)}
\usage{
update_consensus(phi_list, windows, overlapInd)
}
\arguments{
\item{phi_list}{Primal variables}

\item{windows}{Matrix indicating window group}

\item{overlapInd}{Indices of overlap between windows}
}
\description{
\code{update_consensus}
}
