% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_windows.R
\name{update_dual}
\alias{update_dual}
\title{Update dual variable (overlap)}
\usage{
update_dual(w, phi, phiBar, eta, etaBar, rho)
}
\arguments{
\item{w}{dual variable}

\item{phi}{primal variable}

\item{phiBar}{consensus variable}

\item{eta}{primal variable equal to Dtheta}

\item{etaBar}{consensus variable for eta}

\item{rho}{ADMM step size parameter}
}
\description{
\code{update_dual}
}
