% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_updates.R
\name{update_model}
\alias{update_model}
\title{Update model with ADMM step parameter and dual and consensus variables}
\usage{
update_model(model, w, z, rho, nT)
}
\arguments{
\item{model}{List of QP model elements}

\item{w}{dual variable}

\item{z}{Consensus variable}

\item{rho}{ADMM step size}

\item{nT}{Number of quantiles being estimated}
}
\description{
\code{update_model}
}
