% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_updates.R
\name{update_model_abs}
\alias{update_model_abs}
\title{Update model with ADMM step parameter and dual and consensus variables}
\usage{
update_model_abs(model, w, z, nT)
}
\arguments{
\item{model}{Model object}

\item{w}{dual variable}

\item{z}{consensus variable}

\item{nT}{number of quantile levels}
}
\description{
\code{update_model_abs}
}
