% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_checkLoss.R
\name{valid_checkLoss}
\alias{valid_checkLoss}
\title{Calculate total check loss for validation data}
\usage{
valid_checkLoss(lambda, folds, y, tau, k)
}
\arguments{
\item{lambda}{smoothing penalty parameter}

\item{folds}{vector indicating folds for k-fold cross-validation}

\item{y}{observed data}

\item{tau}{quantile levels at which to evaluate trend}

\item{k}{order of differencing}
}
\description{
\code{valid_checkLoss} Calculate checkloss for validation data
}
