% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkloss.R
\name{checkloss}
\alias{checkloss}
\title{Evaluate checkloss function}
\usage{
checkloss(e, tau)
}
\arguments{
\item{e}{argument of checkloss function}

\item{tau}{quantile to be used}
}
\description{
\code{checkloss}
}
