% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{get_Dk}
\alias{get_Dk}
\title{kth order difference matrix}
\usage{
get_Dk(n, k)
}
\arguments{
\item{n}{length of input}

\item{k}{order of the derivative}
}
\description{
\code{get_Dk} computes the discrete kth derivative matrix.
}
\examples{
n <- 10
k <- 3
D <- get_Dk(n, k)
}
