% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_windows_BIC.R
\name{get_windows_BIC}
\alias{get_windows_BIC}
\title{Choose smoothing parameter using SIC}
\usage{
get_windows_BIC(y, tau, k, window_size, overlap,
  lambdaSeq = length(y)^seq(0, 1.5, length.out = 20), df_tol = 1e-09,
  gamma = 1, plot_lambda = FALSE, solver = NULL, criteria = "eBIC")
}
\arguments{
\item{y}{observed data}

\item{tau}{quantile levels at which to evaluate trend}

\item{k}{order of differencing}

\item{lambdaSeq}{smoothing penalty parameter options to compare}

\item{df_tol}{tolerance for determining degrees of freedom (D%*%theta > df_tol)}

\item{plot_lambda}{TRUE/FALSE for plotting lambda by model criteria}

\item{solver}{LP solver, can be "gurobi", "Rglpk", or "lpSolve"}

\item{parameter}{for eBIC}

\item{criter}{criteria to use for lambda selection, must be "eBIC", "SIC", or 
"valid"}
}
\description{
\code{get_windows_BIC} Selects smoothing parameter using Schwarz Information
}
