% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBaseline.R
\name{getBaseline}
\alias{getBaseline}
\title{Get baseline using trend filtering}
\usage{
getBaseline(y, lambda0 = 1e-10, tau = 0.05)
}
\arguments{
\item{y}{vector of numbers to fit baseline}

\item{tau}{quantile level}

\item{lambda}{regularization parameter, larger values produce 
smoother baselines}

\item{maxiter}{number of iterations to use in algorithm}
}
\description{
Get smooth baseline of data. Prints data and baseline estimate every 5000
iterarions. Data may contain missing values. Missing values are replaced using 
next observation carry backward.
}
\examples{

}
