% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{project_V}
\alias{project_V}
\title{Project onto subspace}
\usage{
project_V(theta, eta, D, cholM, k)
}
\arguments{
\item{theta}{first input}

\item{eta}{second input}

\item{D}{differencing matrix}

\item{cholM}{upper triangular cholesky decomposition of  I + DtD}

\item{k}{order of differencing matrix}
}
\description{
\code{project_V} projects (theta, eta) onto the subspace eta = D*theta.
Updates values of theta and eta in place.
}
