% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox.R
\name{prox}
\alias{prox}
\title{Proximal mapping}
\usage{
prox(theta, eta, y, lambda, tau = 0.05, t = 1)
}
\arguments{
\item{theta}{input}

\item{eta}{input}

\item{y}{response}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{t}{step-size}
}
\description{
\code{prox} computes the block separable proximal mapping.
}
\examples{
set.seed(12345)
}
