% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_quantile}
\alias{prox_quantile}
\title{Proximal Mapping}
\usage{
prox_quantile(w, tau, alpha)
}
\arguments{
\item{w}{input}

\item{tau}{quantile parameter}

\item{alpha}{scale parameter}
}
\description{
\code{prox_quantile} computes the proximal mapping of the check function.
}
