% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_functions.R
\name{solve_model}
\alias{solve_model}
\title{Solves the model using chosen solver}
\usage{
solve_model(model, solver, y = NULL, trend = TRUE)
}
\arguments{
\item{y}{observed data, should be equally spaced, may contain NA}

\item{trend}{if TRUE returns trend, if FALSE returns residuals}

\item{tau}{quantile levels at which to evaluate trend}

\item{lambda}{penalty paramter controlling smoothness}

\item{k}{order of differencing}
}
\description{
\code{solve_model} Returns the trend
}
