% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spingarn_multi_step}
\alias{spingarn_multi_step}
\title{Multiple steps of Spingarn's algorithm}
\usage{
spingarn_multi_step(theta, eta, y, D, cholM, lambda, tau = 0.05, step = 1,
  numberIter = 1, k = 3L)
}
\arguments{
\item{theta}{input 1}

\item{eta}{input 2}

\item{y}{response}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{step}{step-size}

\item{numberIter}{number of iterations}

\item{k}{order of derivative}
}
\description{
\code{spingarn_one_step}
}
\examples{
set.seed(12345)
n <- 4e3
x <- seq(1/n, 1, length.out=n)
f <- 2*(x + 2)^2 + 3*cos(3*pi*x)
tau <- 5e4
g1 <- 40*exp(-tau*(x-0.3)^2)
g2 <- 30*exp(-3*tau*(x-0.5)^2)
g3 <- 37*exp(-tau*(x-0.7)^2)
g4 <- 45*exp(-tau/10*(x-0.55)^2)
y <- f + g1 + g2 + g3 + g4 + rnorm(n)
plot(y~x, type="l")
k <- 3
D <- get_Dk(n, k)
M <- diag(n) + crossprod(D)
cholM <- as.matrix(chol(M))
lambda <- 10
tau <- 0.01
step <- 1
numberIter <- 100
multi_step <- spingarn_multi_iter(theta, eta, y, n, k, lambda, 
tau, step, numberIter)
theta <- multi_step[[1]]
theta_last <- prox_f1(theta, y, tau)
plot(x,f,type='l',col='blue', ylim=c(min(y),max(y)), lwd=3)
points(x,y,pch=16)
lines(x,theta_last,col='red', lwd=3)
}
