% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spingarn_one_step}
\alias{spingarn_one_step}
\title{One step of Spingarn's algorithm}
\usage{
spingarn_one_step(theta, eta, Vdiff, y, D, cholM, lambda, tau = 0.05,
  step = 1, k = 3L)
}
\arguments{
\item{theta}{input 1}

\item{eta}{input 2}

\item{y}{response}

\item{D}{differencing matrix}

\item{cholM}{upper cholesky of  (I + DtD)}

\item{lambda}{regularization parameter}

\item{tau}{quantile parameter}

\item{step}{step-size}
}
\description{
\code{spingarn_one_step} updates theta and eta in place
}
