% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_updates.R
\name{update_dual}
\alias{update_dual}
\title{Update dual variable (overlap)}
\usage{
update_dual(w, phi, phiBar, rho)
}
\arguments{
\item{w}{dual variable}

\item{phi}{primal variable}

\item{phiBar}{consensus variable}

\item{rho}{ADMM step size parameter}
}
\description{
\code{update_dual}
}
