% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_windows.R
\name{update_model}
\alias{update_model}
\title{Update model with ADMM step parameter and dual and consensus variables}
\usage{
update_model(model, w, phiBar, etaBar, rho, nT)
}
\arguments{
\item{model}{List of QP model elements}

\item{w}{dual variable}

\item{phiBar}{consensus variable for phi}

\item{etaBar}{consensus variable for eta}

\item{rho}{ADMM step size}

\item{nT}{Number of quantiles being estimated}
}
\description{
\code{update_model}
}
