% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_updates.R
\name{update_windows}
\alias{update_windows}
\title{Update windows}
\usage{
update_windows(y_list, w_list, phiBar_list, tau, lambda, D, rho)
}
\arguments{
\item{rho}{step size for ADMM}
}
\description{
\code{update_windows} Minimizes the Lagrangian for each window
}
