% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_updates.R
\name{update_windows}
\alias{update_windows}
\title{Update windows}
\usage{
update_windows(w_list, phiBar_list, model_list, rho, nT)
}
\arguments{
\item{rho}{step size for ADMM}
}
\description{
\code{update_windows} Minimizes the Lagrangian for each window
}
