% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_windows.R
\name{update_windows}
\alias{update_windows}
\title{Update windows}
\usage{
update_windows(w_list, phiBar_list, etaBar_list, model_list, rho, nT,
  solver = "gurobi")
}
\arguments{
\item{w_list}{Dual variable}

\item{phiBar_list}{Consensus variable for phi}

\item{etaBar_list}{Consensus variable for eta}

\item{model_list}{List of model objects for solver}

\item{rho}{ADMM parameter}

\item{nT}{number of quantiles}

\item{solver}{must be "gurobi" or "quadprog"}
}
\description{
\code{update_windows} Minimizes the Lagrangian for each window
}
