% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmStart.R
\name{warmStart}
\alias{warmStart}
\title{Warm start for Spingarn algorithm}
\usage{
warmStart(y, k, lambda, step, tau, reduction, maxiter)
}
\arguments{
\item{y}{original data to fit baseline}

\item{k}{order of differencing matrix}

\item{lambda}{regularization parameter}

\item{tau}{quantile level}

\item{reduction}{number of observations to aggregate}
}
\description{
Get approximation of initial values by aggregating data and fitting on 
smaller dataset
}
